refactor(mcp): move McpParams/McpResponseTrim to util package, guard RCA error messages
