+++ openmetadata-mcp/src/main/java/org/openmetadata/mcp/util/McpResponseTrim.java
@@ -1,4 +1,4 @@
-package org.openmetadata.mcp.tools;
+package org.openmetadata.mcp.util;
 
 import java.util.LinkedHashMap;
 import java.util.List;
